feat: Add Vercel configuration file
This commit adds a Vercel configuration file to the repository. The file specifies the version of Vercel to use and defines the build and route settings for the deployment. The \builds\ section specifies that the \main.py\ file should be used with the \@vercel/python\ builder. The \routes\ section defines a catch-all route that redirects all requests to the \main.py\ file.